refactor(enums): move all enums to helpers/enums
