+++ src/helpers/enums/activity-level.enum.ts
@@ -1,0 +1,7 @@
+export enum ActivityLevel {
+  SEDENTARY = 'sedentary',
+  LIGHTLY_ACTIVE = 'lightly_active',
+  MODERATELY = 'moderately',
+  ACTIVE = 'active',
+  VERY_ACTIVE = 'very_active',
+}
