+++ src/helpers/enums/activity-type.enum.ts
@@ -1,0 +1,6 @@
+export enum ActivityType {
+  CARDIO = 'cardio',
+  STRENGTH = 'strength',
+  FLEXIBILITY = 'flexibility',
+  BALANCE = 'balance',
+}
