+++ src/helpers/enums/attach-type.enum.ts
@@ -1,0 +1,4 @@
+export enum AttachType {
+  NONE = 'none',
+  IN_APP = 'in_app',
+}
