+++ src/helpers/enums/bfp-calculating-method.enum.ts
@@ -1,0 +1,5 @@
+export enum BFPCalculatingMethod {
+  BMI = 'bmi',
+  US_NAVY = 'us_navy',
+  YMCA = 'ymca',
+}
