+++ src/helpers/enums/challenge-difficulty.enum.ts
@@ -1,0 +1,5 @@
+export enum ChallengeDifficulty {
+  EASY = 'easy',
+  MEDIUM = 'medium',
+  HARD = 'hard',
+}
