+++ src/helpers/enums/chat-session-status.enum.ts
@@ -1,0 +1,4 @@
+export enum ChatSessionStatus {
+  CHAT = 'chat',
+  CONSULT = 'consult',
+}
