+++ src/helpers/enums/diet-type-name.enum.ts
@@ -1,0 +1,7 @@
+export enum DietTypeName {
+  KETO = 'keto',
+  PALEO = 'paleo',
+  VEGAN = 'vegan',
+  MEDITERRANEAN = 'mediterranean',
+  DASH = 'dash',
+}
