+++ src/helpers/enums/fitness-goal-type.enum.ts
@@ -1,0 +1,7 @@
+export enum FitnessGoalType {
+  CUT = 'cut',
+  BULK = 'bulk',
+  MAINTAIN = 'maintain',
+  RECOVERY = 'recovery',
+  GAIN_MUSCLES = 'gain_muscles',
+}
