+++ src/helpers/enums/food-type.enum.ts
