+++ src/helpers/enums/medal-tier.enum.ts
@@ -1,0 +1,6 @@
+export enum MedalTier {
+  BRONZE = 'bronze',
+  SILVER = 'silver',
+  GOLD = 'gold',
+  PLATINUM = 'platinum',
+}
