+++ src/helpers/enums/media-type.enum.ts
@@ -1,0 +1,5 @@
+export enum MediaType {
+  IMAGE = 'image',
+  VIDEO = 'video',
+  FILE = 'file',
+}
