+++ src/helpers/enums/message-type.enum.ts
@@ -1,0 +1,5 @@
+export enum MessageType {
+  TEXT = 'text',
+  IMAGE = 'image',
+  FILE = 'file',
+}
