+++ src/helpers/enums/premium-package-name.enum.ts
@@ -1,0 +1,5 @@
+export enum PremiumPackageName {
+  FREE = 'free',
+  LITE = 'lite',
+  PRO = 'pro',
+}
