+++ src/helpers/enums/record-type.enum.ts
@@ -1,0 +1,5 @@
+export enum RecordType {
+  DAILY = 'daily',
+  GOAL = 'goal',
+  CHALLENGE = 'challenge',
+}
