+++ src/helpers/transformers/money.transformer.ts
@@ -1,0 +1,5 @@
+// hàm đưa và trả tiền ra vào db
+export const Money = {
+  to: (v?: number | null) => v,
+  from: (v?: string | null) => (v == null ? null : Number(v)),
+};
