+++ src/premium_packages/entities/premium_package.entity.ts
@@ -1,18 +1,8 @@
 import { ApiProperty, ApiSchema } from '@nestjs/swagger';
 import { User } from 'src/users/entities/user.entity';
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn, UpdateDateColumn } from 'typeorm';
-
-export enum PremiumPackageName {
-  FREE = 'free',
-  LITE = 'lite',
-  PRO = 'pro',
-}
-
-// hàm đưa và trả tiền ra vào db
-export const Money = {
-  to: (v?: number | null) => v,
-  from: (v?: string | null) => (v == null ? null : Number(v)),
-};
+import { PremiumPackageName } from 'src/helpers/enums/premium-package-name.enum';
+import { Money } from 'src/helpers/transformers/money.transformer';
 
 @ApiSchema({ name: PremiumPackage.name, description: 'PremiumPackage entity' })
 @Entity('premium_packages')
